+++ src/index.ts
@@ -4,4 +4,18 @@ export * from './structs/coercions.js';
 export * from './structs/refinements.js';
 export * from './structs/types.js';
 export * from './structs/utilities.js';
-export type { ObjectSchema, Optionalize, Simplify } from './utils.js';
+export type {
+  AnyStruct,
+  EnumSchema,
+  InferStructTuple,
+  IsExactMatch,
+  IsMatch,
+  IsRecord,
+  IsTuple,
+  ObjectSchema,
+  OmitBy,
+  Optionalize,
+  PickBy,
+  Simplify,
+  UnionToIntersection,
+} from './utils.js';